requirements.txt
fastapi
strawberry-graphql
dishka
uvicorn